+++ freemius/templates/forms/license-activation.php
@@ -52,9 +52,8 @@
     if ( $is_network_activation ) {
         $all_sites = Freemius::get_sites();
 
-        $all_site_details          = array();
-        $subsite_url_by_install_id = array();
-        $install_url_by_install_id = array();
+        $subsite_data_by_install_id = array();
+        $install_url_by_install_id  = array();
 
         foreach ( $all_sites as $site ) {
             $site_details = $fs->get_site_info( $site );
@@ -67,9 +66,9 @@
             $install = $fs->get_install_by_blog_id($blog_id);
 
             if ( is_object( $install ) ) {
-                if ( isset( $subsite_url_by_install_id[ $install->id ] ) ) {
-                    $clone_subsite_url = $subsite_url_by_install_id[ $install->id ];
-                    $clone_install_url = $install_url_by_install_id[ $install->id ];
+                if ( isset( $subsite_data_by_install_id[ $install->id ] ) ) {
+                    $clone_subsite_data = $subsite_data_by_install_id[ $install->id ];
+                    $clone_install_url  = $install_url_by_install_id[ $install->id ];
 
                     if (
                         /**
@@ -78,7 +77,7 @@
                          * @author Leo Fajardo (@leorw)
                          * @since 2.5.0
                          */
-                        fs_strip_url_protocol( untrailingslashit( $clone_install_url ) ) === fs_strip_url_protocol( untrailingslashit( $clone_subsite_url ) ) ||
+                        fs_strip_url_protocol( untrailingslashit( $clone_install_url ) ) === fs_strip_url_protocol( untrailingslashit( $clone_subsite_data['url'] ) ) ||
                         fs_strip_url_protocol( untrailingslashit( $install->url ) ) !== fs_strip_url_protocol( untrailingslashit( $site_details['url'] ) )
                     ) {
                         continue;
@@ -89,17 +88,15 @@
                     $site_details['license_id'] = $install->license_id;
                 }
 
-                $subsite_url_by_install_id[ $install->id ] = $site_details['url'];
-                $install_url_by_install_id[ $install->id ] = $install->url;
+                $subsite_data_by_install_id[ $install->id ] = $site_details;
+                $install_url_by_install_id[ $install->id ]  = $install->url;
             }
-
-            $all_site_details[] = $site_details;
         }
 
         if ( $is_network_activation ) {
             $vars = array(
                 'id'                  => $fs->get_id(),
-                'sites'               => $all_site_details,
+                'sites'               => array_values( $subsite_data_by_install_id ),
                 'require_license_key' => true
             );