+++ views/logged-in.php
@@ -72,36 +72,7 @@ if (current_user_can('administrator') && 'on' === $manually_register && is_page(
 				</label>
 				<input type="text" name="wpfep-website" id="wpfep-user_website" class="input" value="<?php echo esc_html($register_obj->get_post_value('wpfep-website')); ?>"  />
 			</li>
-			<li class="wpfep-form-field wpfep-default-user-role">
-				<label for="wpfep-default-role" class="wpfep-default-role">
-				<input type="radio" class="wpfep-showform-default" value="Default role" name="showform" checked="" /><?php esc_attr_e('Default role', 'wpfep'); ?></label>
-				<?php
-                $roles_obj         = new WP_Roles();
-    $roles_names_array = $roles_obj->get_names(); ?>
-				<select name="role" id="wpfep-custom" class="input" style="display:none">
-				<?php
-                foreach ($roles_names_array as $key => $value) {
-                    if (in_array($key, array( 'subscriber' ))) {
-                        echo '<option value="' . esc_html($key) . '">' . esc_html($value) . '</option>';
-                    }
-                } ?>
-				</select>
-				<label for="wpfep-custom-role" class="wpfep-custom-role">
-				<input type="radio" class="wpfep-showform-custom" value="Custom role" name="showform" class="wpfep-role-right"/><?php esc_attr_e('Custom role', 'wpfep'); ?></label><br><br>
-				<div id="wpfep-default"></div>
-				<?php
-                $roles_obj         = new WP_Roles();
-    $roles_names_array = $roles_obj->get_names(); ?>
-				<select name="role" id="wfp-user-role" class="input" style="display:none">
-					<option value="" disabled selected><?php esc_attr_e('Select user role', 'wpfep'); ?></option>
-					<?php
-                    foreach ($roles_names_array as $key => $value) {
-                        if (in_array($key, array( 'editor', 'author', 'contributor' ))) {
-                            echo '<option value="' . esc_html($key) . '">' . esc_html($value) . '</option>';
-                        }
-                    } ?>
-				</select>
-			</li>
+			
 			<li class="wpfep-form-field wpfep-default-user-bio">
 				<label for="wpfep-description"><?php esc_attr_e('Biographical Info', 'wpfep'); ?>
 				</label>