+++ wp-frontend-profile.php
@@ -3,7 +3,7 @@
  * Plugin Name: WP Frontend Profile
  * Plugin URI: https://wordpress.org/plugins/wp-front-end-profile/
  * Description: This plugin allows users to easily edit their profile information on the frontend rather than having to go into the dashboard to make changes to password, email address and other user meta data.
- * Version:     1.3.5
+ * Version:     1.3.6
  * @package wp-front-end-profile
  * Author:      Glowlogix
  * Author URI:  https://www.glowlogix.com
@@ -17,7 +17,7 @@ defined('ABSPATH') || exit;
  * Main class for WP Frontend Profile.
  */
 if (! defined('WPFEP_VERSION')) {
-    define('WPFEP_VERSION', '1.3.5');
+    define('WPFEP_VERSION', '1.3.6');
 }
 if (! defined('WPFEP_PATH')) {
     define('WPFEP_PATH', plugin_dir_path(__FILE__));
@@ -29,9 +29,7 @@ if (! defined('WPFEP_PLUGIN_URL')) {
 require_once WPFEP_PATH . '/inc/class-wp-frontend-profile.php';
 
 if (! function_exists('wfep_fs')) {
-    /**
-     * Create a helper function for easy SDK access.
-     */
+    // Create a helper function for easy SDK access.
     function wfep_fs()
     {
         global $wfep_fs;
@@ -40,29 +38,27 @@ if (! function_exists('wfep_fs')) {
             // Include Freemius SDK.
             require_once dirname(__FILE__) . '/freemius/start.php';
 
-            $wfep_fs = fs_dynamic_init(
-                array(
-                    'id'                  => '5837',
-                    'slug'                => 'wp-front-end-profile',
-                    'premium_slug'        => 'wp-frontend-profile-premium',
-                    'type'                => 'plugin',
-                    'public_key'          => 'pk_ac83abfabd6c3c1498e82893f4a23',
-                    'is_premium'          => true,
-                    // If your plugin is a serviceware, set this option to false.
-                    'has_premium_version' => true,
-                    'has_addons'          => true,
-                    'has_paid_plans'      => true,
-                    'trial'               => array(
-                        'days'               => 7,
-                        'is_require_payment' => false,
-                    ),
-                    'menu'                => array(
-                        'slug'    => 'wpfep-settings_dashboard',
-                        'support' => false,
-                    ),
-
-                )
-            );
+            $wfep_fs = fs_dynamic_init(array(
+                'id'                  => '5837',
+                'slug'                => 'wp-front-end-profile',
+                'premium_slug'        => 'wp-frontend-profile-premium',
+                'type'                => 'plugin',
+                'public_key'          => 'pk_ac83abfabd6c3c1498e82893f4a23',
+                'is_premium'          => true,
+                // If your plugin is a serviceware, set this option to false.
+                'has_premium_version' => true,
+                'has_addons'          => false,
+                'has_paid_plans'      => true,
+                'trial'               => array(
+                    'days'               => 14,
+                    'is_require_payment' => false,
+                ),
+                'has_affiliation'     => 'selected',
+                'menu'                => array(
+                    'slug'           => 'wpfep-settings_dashboard',
+                    'support'        => false,
+                ),
+            ));
         }
 
         return $wfep_fs;