+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file, per [the Ke
 
 ## [1.3.6] - 2024-07-19
 - Tested with WordPress version 6.7 #138
+- User can be register with different roles #127
 - Update Fremius SDK #142
 - Fix WPFP profile page rendering Issue. #141